requirements.txt
# 示例 requirements.txt
requests
PySocks
telethon
beautifulsoup4
python-dotenv
dnspython